fixed student model save method
1.Error: "Reverse for 'staff-detail' with keyword arguments '{'pk': ''}' not found."

This error occurs when a student, logged in, tries to save data as a staff member. The system can't find the 'staff' model, leading to a missing 'pk' (primary key).

2.Error: "Exception Value: User has no staff_profile."

This happens when the system treats a student as staff, attempting to find a 'staff_profile' relationship for them, which doesn't exist.